+++ src/h.js
@@ -425,21 +425,21 @@ var H = (function() {
     elt.classList[dir === 'r' ? 'remove' : 'add'](cla);
   };
 
-  var rearg_sta_sel_nam_bof = function(args, bofdef) {
+  var rearg_sta_sel_nam_las = function(args, las) {
 
     var a = args[0], b = args[1], c = args[2], d = args[3];
 
-    if (args.length < 2) throw "rearg_sta_sel_nam_bof() not enough arguments";
+    if (args.length < 2) throw "at least 2 arguments required";
 
-    if (args.length === 2) return { sta: a, sel: null, nam: b, bof: bofdef };
-    if (args.length > 3) return { sta: a, sel: b, nam: c, bof: d };
+    if (args.length === 2) return { sta: a, sel: null, nam: b, las: las };
+    if (args.length > 3) return { sta: a, sel: b, nam: c, las: d };
 
-    // sta/sel/nam or sta/nam/bof ?
+    // sta/sel/nam or sta/nam/las ?
 
     if ((typeof c) === 'string' && c.match(/^\.?[^ ]+$/))
-      return { sta: a, sel: b, nam: c, bof: bofdef };
+      return { sta: a, sel: b, nam: c, las: las };
 
-    return { sta: a, sel: null, nam: b, bof: c };
+    return { sta: a, sel: null, nam: b, las: c };
   };
 
   var toggle = function(start, sel, cla, bof, mod) {
@@ -456,12 +456,12 @@ var H = (function() {
   };
 
   this.addClass = function(start, sel, cla, bof) {
-    var as = rearg_sta_sel_nam_bof(arguments, true);
-    toggle(as.sta, as.sel, as.nam, as.bof, 'a');
+    var as = rearg_sta_sel_nam_las(arguments, true);
+    toggle(as.sta, as.sel, as.nam, as.las, 'a');
   }
   this.removeClass = function(start, sel, cla, bof) {
-    var as = rearg_sta_sel_nam_bof(arguments, true);
-    toggle(as.sta, as.sel, as.nam, as.bof, 'r');
+    var as = rearg_sta_sel_nam_las(arguments, true);
+    toggle(as.sta, as.sel, as.nam, as.las, 'r');
   };
 
   this.toggleClass = function(start, sel, cla) {
@@ -474,8 +474,8 @@ var H = (function() {
   this.toggle = this.toggleClass;
 
   this.setClass = function(start, sel, cla, bof) {
-    var as = rearg_sta_sel_nam_bof(arguments, true);
-    toggle(as.sta, as.sel, as.nam, as.bof, 'ar');
+    var as = rearg_sta_sel_nam_las(arguments, true);
+    toggle(as.sta, as.sel, as.nam, as.las, 'ar');
   };
 
   this.renameClass = function(start, sel, cla0, cla1) {
@@ -569,12 +569,22 @@ var H = (function() {
 
   this.getAtt = function(start, sel, aname/*, default*/) {
 
-    var as = rearg_sta_sel_nam_bof(arguments, undefined);
+    var as = rearg_sta_sel_nam_las(arguments, undefined);
 
     var e = H.elt(as.sta, as.sel);
     if ( ! e) throw "elt not found, cannot read attributes";
 
-    return e.getAttribute(as.nam) || as.bof;
+    return e.getAttribute(as.nam) || as.las;
+  };
+
+  this.getAtti = function(start, sel, aname/*, default*/) {
+    var v = self.getAtt.apply(null, arguments);
+    return v ? parseInt(v, 10) : v;
+  };
+
+  this.getAttf = function(start, sel, aname/*, default*/) {
+    var v = self.getAtt.apply(null, arguments);
+    return v ? parseFloat(v) : v;
   };
 
   this.capitalize = function(s) {